[front] refactor: Pass the workspace to the grant cache read path
